docs(router): add worker selection API guidance
